Move typing to a package.
